moved code into player, arena, dice
