+++ sample.txt
@@ -1,0 +1,150 @@
+### Observations and notes ###
+# Based on 80,000 matches per #
+# The main principle of these tests is: "all things being equal" #
+# This helps with a direct comparison without adding other complications #
+
+Level 10 Fighters, Greatswords, Extra Attack, 103HP, 15AC
+    +1 init, 51.5%-48.5% (+3%)
+    +2 init, 52.4%-47.6% (+4.8%)
+    +3 init, 53.1%-46.9% (+6.3%)
+    +4 init, 53.6%-46.4% (+7.2%)
+    +5 init, 54.6%-45.4% (+9.2%)
+    ...
+    +10 init, 57.2%-42.8% (+14.4%)
+
+19AC vs. 18AC resulted in 60%-40% winrate. (+20%)
+20AC vs. 18AC resulted in 70%-30% winrate. (+30%)
+21AC vs. 18AC resulted in 78%-22% winrate in favor. (+56%)
+22AC vs. 18AC resulted in 86%-14% winrate in favor. (+72%)
+23AC vs. 18AC resulted in 92%-8% winrate in favor. (+84%)
+24AC vs. 18AC resulted in 96%-4% winrate in favor. (+92%)
+
+18AC vs. 17AC resulted in 60%-40% winrate. (+20%)
+18AC vs. 16AC resulted in 68%-32% winrate. (+36%)
+18AC vs. 15AC resulted in 75%-25% winrate. (+50%)
+18AC vs. 14AC resulted in 81%-19% winrate. (+62%)
+18AC vs. 13AC resulted in 87%-13% winrate. (+74%)
+18AC vs. 12AC resulted in 90%-10% winrate. (+80%)
+
+Crits on 19 resulted in 52.5%-47.5% winrate. (+5%)
+Crits on 18 resulted in 55%-45% winrate. (+10%)
+
+Greatsword vs. Greataxe resulted in 53%-47% in favor of Greatsword. (+6%)
+
+All things being equal, having Extra Attack vs. not resulted in 89%-11% in favor. (+78%)
+
+Without extra attack:
+    Greatsword vs. dual Shortswords (no ability mod on offhand)
+        49.7%-50.3% in favor of Shortswords
+    Greatsword vs. dual Shortswords with Two-Weapon Fighting
+        20.3%-79.7% in favor of Shortswords
+    Greatsword with Defense Style (+1AC) vs. dual Shortswords with Two-Weapon Fighting
+        28.2%-71.8% in favor of Shortswords
+With extra attack:
+    Greatsword vs. dual Shortswords (no ability mod on offhand)
+        65.6%-34.4% in favor of Greatsword
+    Greatsword vs. dual Shortswords with Two-Weapon Fighting
+        45.2%-54.8% in favor of Shortswords
+    Greatsword with Defense Style (+1AC) vs. dual Shortswords with Two-Weapon Fighting
+        55%-45% in favor of Greatsword
+With extra attack (x2):
+    Greatsword vs. dual Shortswords (no ability mod on offhand)
+        70.6%-29.4% in favor of Greatsword
+    Greatsword vs. dual Shortswords with Two-Weapon Fighting
+        56.7-43.3% in favor of Greatsword
+    Greatsword with Defense Style (+1AC) vs. dual Shortswords with Two-Weapon Fighting
+        65.8%-34.2% in favor of Greatsword
+
+Dual-Wielder Feat (two Rapiers and +1AC) vs. not:
+    71.8%-28.2% in favor of Dual-Wielder Feat (no extra attack, 19vs18 AC)
+    71.6%-28.4% in favor of Dual-Wielder feat (extra attack, 17vs16 AC)
+
+Dual-Wielder Feat (two Rapiers and +1AC) vs. +2 Dex:
+    40.9%-59.1% in favor of +2 Dex, not counting the ability to draw two weapons at once. (+18%)
+
+Dual-Wielder Feat (two Rapiers and +1AC) vs. Heavy Armor Master:
+    29%-71% in favor of Heavy Armor Master. (Level 8, 20 Str/Dex, 2x Extra Attack (4 attacks total), TWF style) (+42%)
+
+Heavy Armor Master vs. not:
+    72%-28% in favor of HAM. (level 8, Extra Attack, Greatswords, 20 Str) (+44%)
+    87%-13% in favor of HAM. (level 8, Extra Attack, dual Shortswords, 20 Str) (+74%)
+
+GWM/SS vs. not:
+    Level 10 Fighters, Extra Attacks, Greataxes, 18 Strength (+8 to attack bonus)
+         9AC: 82%-18% (+64%)
+        10AC: 79%-21% (+58%)
+        11AC: 77%-23% (+54%)
+        12AC: 74%-26% (+48%)
+        13AC: 71%-29% (+42%)
+        14AC: 68%-32% (+36%)
+        15AC: 65%-35% (+30%)
+        16AC: 61%-39% (+22%)
+        17AC: 57%-43% (+14%)
+        18AC: 52%-48% (+4%)
+        19AC: 46%-54% (-8%) //// @19AC it's better to not use GWM/SS
+        20AC: 40%-60% (-20%)
+        21AC: 32%-68% (-36%)
+        22AC: 22%-78% (-56%)
+        23AC: 10%-90% (-80%)
+
+GWM/SS vs. +2 Strength:
+    Level 10 Fighters, Extra Attacks, Greataxes, 18/20 Strength
+        9AC: 75%-25% (+50%)
+        10AC: 68%-32% (+36%)
+        11AC: 66%-34% (+32%)
+        12AC: 62%-38% (+24%)
+        13AC: 60%-40% (+20%)
+        14AC: 57%-43% (+14%)
+        15AC: 53%-47% (+6%)
+        16AC: 49%-51% (-1%) //// @16AC +2 str becomes better.
+        17AC: 45%-55% (-10%)
+        18AC: 41%-59% (-18%)
+        19AC: 35%-65% (-30%)
+        20AC: 29%-71% (-42%)
+        21AC: 22%-78% (-56%)
+        22AC: 14%-86% (-72%)
+        23AC: 06%-94% (-88%)
+
+GWM/SS vs. +2 Strength:
+    Level 5 Fighters, Extra Attacks, Greatswords, 18/20 Strength
+        ...
+        14AC: 50.5%-49.5% (+1%)
+        15AC: 48%-52% (-4%) //// @15AC +2 str becomes better.
+        ...
+
+155HP vs. 150HP resulted in 54%-46% winrate in favor. (+8%)
+105HP vs. 100HP resulted in 55%-45% winrate in favor. (+10%)
+55HP vs. 50HP resulted in 57%-43% winrate in favor. (+14%)
+25HP vs. 20HP resulted in 60%-40% winrate in favor. (+20%)
+15HP vs. 10HP resulted in 65%-35% winrate in favor. (+30%)
+
+### Constitution>HP tests ###
+
+Level 1 Barbarian
+    +1 Conmod: 53%-47% (+6%)
+    +2 Conmod: 56%-44% (+12%)
+    +3 Conmod: 59%-41% (+18%)
+    +4 Conmod: 62%-38% (+24%)
+    +5 Conmod: 66%-34% (+32%)
+
+Level 2 Barbarian
+    +1 Conmod: 56%-43%
+    +2 Conmod: 63%-37%
+    +3 Conmod: 68%-31%
+    +4 Conmod: 73%-27%
+    +5 Conmod: 78%-22%
+
+Level 5 Barbarian
+    +1 Conmod: 59.5%-40.5% (+19%)
+    ...
+    +5 Conmod: 87%-13% (+74%)
+
+Level 10 Barbarian
+    +1 Conmod: 64%-36% (+28%)
+    ...
+    +5 Conmod: 94%-6% (+88%)
+
+Level 20 Barbarian
+    +1 Conmod: 69%-31% (+38%)
+    ...
+    +5 Conmod: 98.8%-1.2% (+97.6%)
